loan,migrations

+++ src/main/resources/db/migration/V1__initUsers.sql
@@ -1,5 +1,5 @@
 create table users (
-    id int primary key AUTO_INCREMENT,
+    id bigint primary key AUTO_INCREMENT,
     first_name varchar(100),
     last_name varchar(100) not null,
     email varchar(100) not null
